requirements.txt
accelerate
transformers
hydra-core
slider @ git+https://github.com/OliBomby/slider.git@gedagedigedagedaoh
wandb
pandas
pyarrow
matplotlib
pytest
